refactor: replace hardcode sendEchoMessage method in py websocket client
